requirements.txt
setuptools>=65.5.1
wheel>=0.40.0
numpy>=1.26.0
opencv-python>=4.8.1
tensorflow-cpu>=2.15.0
keras>=2.15.0
pillow>=10.1.0
scikit-learn>=1.3.2
pandas>=2.1.3 